🎨 Style: SkillStack 컴포넌트 분리
Style: SkillStack 컴포넌트 분리
